wifi: mt76: mt7915: use nss for calculating rx-chains
When group-5 is enabled, we can find the actuall NSS used.
In that case, update the rx chains info so that the upper
stack can give better idea of actual antenna usage.

In case of stbc, the hardware already double nss, but rssi
is not reported properly (it is '17') for the secondary chains.
So, decrease nss in the status msg so that mac80211 does not try
to average the invalid RSSI.

Signed-off-by: Ben Greear <greearb@candelatech.com>